[Misc] centralize reference implementations in naive.py
