+++ fla/ops/attn/__init__.py
@@ -1,6 +1,8 @@
 
+from .naive import naive_parallel_attn
 from .parallel import parallel_attn
 
 __all__ = [
+    'naive_parallel_attn',
     'parallel_attn',
 ]
