+++ fla/ops/comba/__init__.py
@@ -1,7 +1,10 @@
 from .chunk import chunk_comba
 from .fused_recurrent import fused_recurrent_comba
+from .naive import naive_chunk_comba, naive_recurrent_comba
 
 __all__ = [
     "chunk_comba",
     "fused_recurrent_comba",
+    "naive_chunk_comba",
+    "naive_recurrent_comba",
 ]
