+++ fla/ops/forgetting_attn/__init__.py
@@ -1,6 +1,8 @@
 
+from .naive import naive_forgetting_attn
 from .parallel import parallel_forgetting_attn
 
 __all__ = [
+    'naive_forgetting_attn',
     'parallel_forgetting_attn',
 ]
