+++ fla/ops/gated_delta_rule/__init__.py
@@ -1,7 +1,10 @@
 from .chunk import chunk_gated_delta_rule
 from .fused_recurrent import fused_recurrent_gated_delta_rule
+from .naive import naive_chunk_gated_delta_rule, naive_recurrent_gated_delta_rule
 
 __all__ = [
     "chunk_gated_delta_rule",
     "fused_recurrent_gated_delta_rule",
+    "naive_chunk_gated_delta_rule",
+    "naive_recurrent_gated_delta_rule",
 ]
