+++ fla/ops/path_attn/__init__.py
@@ -1,6 +1,8 @@
 
+from .naive import naive_path_attn
 from .parallel import parallel_path_attn
 
 __all__ = [
+    'naive_path_attn',
     'parallel_path_attn',
 ]
