+++ tests/ops/test_comba.py
@@ -5,9 +5,9 @@ import os
 import pytest
 import torch
 import torch.nn.functional as F
-from einops import rearrange
 
 from fla.ops.comba import chunk_comba, fused_recurrent_comba
+from fla.ops.comba.naive import naive_chunk_comba
 from fla.ops.comba.utils import chunk_comba_cumsum_scalar_fwd
 from fla.utils import IS_INTEL_ALCHEMIST, assert_close, device
 
@@ -50,85 +50,6 @@ def test_cumsum_local_scalar_fwd(
     assert_close("local cumsum scalar", ref_1, tri_1, 0.001 if dtype == torch.float else 0.003)
 
 
-def chunk_comba_ref(
-    q: torch.Tensor,
-    k: torch.Tensor,
-    v: torch.Tensor,
-    p: torch.Tensor,
-    g: torch.Tensor,
-    beta: torch.Tensor,
-    chunk_size: int = 64,
-    scale: float = None,
-    initial_state: torch.Tensor = None,
-    output_final_state: bool = False,
-):
-    BT = chunk_size
-    if scale is None:
-        scale = 1 / (q.shape[-1] ** 0.5)
-    # Calculate padding needed to make T a multiple of BT
-    q, k, v, p, beta, g = map(lambda x: x.transpose(1, 2).contiguous().to(torch.float32), [q, k, v, p, beta, g])
-
-    T = q.shape[-2]
-    pad_len = (BT - (T % BT)) % BT
-    if pad_len > 0:
-        # Pad all tensors
-        q = F.pad(q, (0, 0, 0, pad_len))
-        k = F.pad(k, (0, 0, 0, pad_len))
-        v = F.pad(v, (0, 0, 0, pad_len))
-        p = F.pad(p, (0, 0, 0, pad_len))
-        beta = F.pad(beta, (0, pad_len))
-        g = F.pad(g, (0, pad_len))
-    decay = g
-    chunk_size = BT
-    b, h, l, d_k = q.shape
-    d_v = v.shape[-1]
-    q = q * scale
-    v = v * beta[..., None]
-    p_beta = p * beta[..., None]
-    assert l % chunk_size == 0
-    # note that diagonal is masked.
-    mask = torch.triu(torch.ones(chunk_size, chunk_size, dtype=torch.bool, device=q.device), diagonal=0)
-    q, k, v, p_beta, decay, g = map(
-        lambda x: rearrange(x, 'b h (n c) d -> b h n c d', c=chunk_size),
-        [q, k, v, p_beta, decay.unsqueeze(-1), g.unsqueeze(-1)],
-    )
-    decay = decay.squeeze(-1).cumsum(-1)  # [B, H, n, c]
-    decay_0 = decay - g.squeeze(-1)  # [B, H, n, c]
-    L_mask = ((decay.unsqueeze(-1) - decay.unsqueeze(-2)).tril().exp().float()).tril()
-    L_mask_0 = ((decay_0.unsqueeze(-1) - decay.unsqueeze(-2)).tril().exp().float()).tril()
-    # [B, H, n, c, d] @ [B, H, n, d, c] -> [B, H, n, c, c]
-    attn = -((p_beta @ k.transpose(-1, -2)) * L_mask_0).masked_fill(mask, 0)
-    for i in range(1, chunk_size):
-        attn[..., i, :i] = attn[..., i, :i].clone() + (attn[..., i, :i, None].clone() * attn[..., :i, :i].clone()).sum(-2)
-    attn = attn + torch.eye(chunk_size, dtype=torch.float, device=q.device)
-    # for U
-    k_cumsum = attn @ v
-    # for W
-    k_cumdecay = attn @ (p_beta * decay_0[..., None].exp())
-    v = k_cumsum
-    S = k.new_zeros(b, h, d_k, d_v)
-    if initial_state is not None:
-        S += initial_state
-    o = torch.zeros_like(v)
-    mask = torch.triu(torch.ones(chunk_size, chunk_size, dtype=torch.bool, device=q.device), diagonal=1)
-    for i in range(0, l // chunk_size):
-        q_i, k_i, v_i = q[:, :, i], k[:, :, i], v[:, :, i]
-        attn = (q_i @ k_i.transpose(-1, -2) * L_mask[:, :, i]).masked_fill_(mask, 0)
-        v_prime = k_cumdecay[:, :, i] @ S
-        v_new = v_i - v_prime
-        o_inter = (q_i * decay[:, :, i, :, None].exp()) @ S
-        o[:, :, i] = o_inter + attn @ v_new
-        S = S * decay[:, :, i, -1, None, None].exp() + (k_i * (decay[:, :, i, -1, None] - decay[:, :, i]).exp()
-                                                        [..., None]).transpose(-1, -2) @ v_new
-    if not output_final_state:
-        S = None
-    # unpad
-    o = rearrange(o, 'b h n c d -> b h (n c) d')
-    o = o[:, :, :T]
-    o = o.transpose(1, 2)
-    return o, S
-
-
 @pytest.mark.parametrize(
     ('B', 'T', 'H', 'D', 'scale', 'gate_logit_normalizer', 'dtype'),
     [
@@ -164,7 +85,7 @@ def test_fused_recurrent(
     g = g / gate_logit_normalizer
     h0 = torch.randn(B, H, D, D, dtype=torch.float32)
     q, k, v, p, beta, g, h0 = map(lambda x: x.to(device).requires_grad_(), (q, k, v, p, beta, g, h0))
-    ref, ref_ht = chunk_comba_ref(
+    ref, ref_ht = naive_chunk_comba(
         q=q.clone(),
         k=k.clone(),
         v=v.clone(),
@@ -252,7 +173,7 @@ def test_chunk(
     tri_dq, tri_dk, tri_dv, tri_dp, tri_dbeta, tri_dg, tri_dh0 = q.grad, k.grad, v.grad, p.grad, beta.grad, g.grad, h0.grad
     q.grad = k.grad = v.grad = p.grad = beta.grad = g.grad = h0.grad = None
 
-    ref, ref_ht = chunk_comba_ref(
+    ref, ref_ht = naive_chunk_comba(
         q=F.normalize(q.clone(), p=2, dim=-1),
         k=F.normalize(k.clone(), p=2, dim=-1),
         p=F.normalize(p.clone(), p=2, dim=-1),
@@ -341,7 +262,7 @@ def test_chunk_varlen(
     ref = []
     ref_ht = []
     for i in range(N):
-        ref_i, ref_ht_i = chunk_comba_ref(
+        ref_i, ref_ht_i = naive_chunk_comba(
             q=q[:, cu_seqlens[i]:cu_seqlens[i+1]],
             k=k[:, cu_seqlens[i]:cu_seqlens[i+1]],
             v=v[:, cu_seqlens[i]:cu_seqlens[i+1]],
